# BaseDatos/openapi-gateway.yaml
swagger: "2.0"
info:
  title: "basedatos"
  version: "1.0.0"
schemes:
  - https
consumes:
  - application/json
produces:
  - application/json
securityDefinitions:
  api_key:
    type: apiKey
    name: x-api-key
    in: header
security:
  - api_key: []

# Backend global: evita repetirlo en cada path
...
paths:
...
  /corregir_ejercicio:
    options:
      operationId: corregir_ejercicio_options
      security: []
      responses:
        "200":
          description: CORS preflight
    post:
      operationId: corregir_ejercicio
      parameters:
        - name: body
          in: body
          required: true
          schema:
            type: object
            required: [alumno_id, id_ejercicio, imagen_base64]
            properties:
              centro_id:
                type: string
              alumno_id:
                type: string
              clase:
                type: string
              asignatura:
                type: string
              unidad:
                type: integer
              id_ejercicio:
                type: string
              imagen_base64:
                type: string
      responses:
        "200":
          description: OK
...